requirements.txt
absl-py==0.9.0
altgraph==0.17
appdirs==1.4.4
arrayfire==3.6.20181017
arrow==0.17.0
asgiref==3.2.7
astropy==4.0
astroquery==0.4
astunparse==1.6.3
atomicwrites==1.3.0
atrader==3.2.2.13
attrs==19.3.0
audioread==2.1.8
Automat==20.2.0
autopep8==1.5.4
backcall==0.1.0
baidu-aip==2.2.18.0
beautiful==0.0.2
beautifulsoup4==4.8.2
bleach==3.1.3
blurhash==1.1.4
boost==0.1
boto==2.49.0
boto3==1.14.21
botocore==1.17.21
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
cmake==3.18.4.post1
cocos==0.1.8
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupy==7.4.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
Django==3.0.5
dlib==19.21.0
docopt==0.6.2
docutils==0.15.2
dukpy==0.2.2
echarts-china-cities-pypkg==0.0.9
echarts-china-counties-pypkg==0.0.2
echarts-china-misc-pypkg==0.0.1
echarts-china-provinces-pypkg==0.0.3
echarts-cities-pypkg==0.0.1
echarts-countries-pypkg==0.1.6
echarts-themes-pypkg==0.0.3
entrypoints==0.3
enum34==1.1.10
et-xmlfile==1.0.1
fastrlock==0.4
ffmpeg==1.4
future==0.18.2
gast==0.3.3
gensim==3.8.3
gm==3.0.127
gmapi==1.17.30
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
imageio==2.8.0
incremental==17.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
javascripthon==0.11
jdcal==1.4.1
jedi==0.16.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
json5==0.9.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-echarts-pypkg==0.1.2
jupyter-lsp==0.8.0
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kaggle==1.5.6
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.2.0
kiwisolver==1.1.0
lab==5.5
librosa==0.8.0
llvmlite==0.34.0
lml==0.0.2
logzero==1.5.0
loky==2.6.0
lxml==4.5.0
macropy3==1.1.0b2
Markdown==3.2.1
MarkupSafe==1.1.1
Mastodon.py==1.5.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
mpmath==1.1.0
multiprocess==0.70.9
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.5
nodejs==0.1.1
notebook==6.0.3
npm==0.1.1
numba==0.51.2
numpy==1.18.2
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.2.1
optional-django==0.1.0
packaging==20.3
pandas==1.0.3
pandoc-fignos==2.2.0
pandoc-xnos==2.3.0
pandocfilters==1.4.2
parsel==1.5.2
parso==0.6.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
pipreqs==0.4.10
plotly==4.7.1
pluggy==0.13.1
ply==3.11
pooch==1.2.0
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
Protego==0.1.16
protobuf==3.13.0
psutil==5.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pygame==1.9.6
Pygments==2.6.1
PyHamcrest==2.0.2
PyInstaller==3.6
pymediainfo==4.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyrsistent==0.15.7
pytest==5.4.1
python-dateutil==2.8.1
python-docx==0.8.10
python-magic==0.4.18
python-slugify==4.0.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
queuelib==1.5.0
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
schedule==0.6.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.0.1
seaborn==0.10.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smart-open==2.1.0
SoundFile==0.10.3.post1
soup==0.1.0
soupsieve==2.0
SQLAlchemy==1.3.20
sqlparse==0.3.1
sympy==1.5.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
tushare==1.2.61
Twisted==20.3.0
txt2tags==3.7
typing==3.7.4.3
urllib3==1.24.3
w3lib==1.21.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.12.1
wxPython==4.0.7.post2
xlrd==1.2.0
yarg==0.1.9
zope.interface==5.0.1
